feat(a2ui-hello-world): add copilotkit.a2ui Python SDK module and a2ui_operations container
- copilotkit/a2ui.py: load_schema(), surface_update(), data_model_update(),
  begin_rendering(), render() — canonical SDK for building A2UI operations
- A2UIMessageRenderer reads a2ui_operations (with operations fallback)
- local-install script now includes editable pip install of Python SDK
- search_flights uses `from copilotkit import a2ui` pattern
